тестовые данные для состава смеси и номенклатуры

+++ database/migrations/2023_03_22_153215_create_mixture_composition_items_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('mixture_composition_id')->constrained();
             $table->foreignId('nomenclature_id')->constrained();
             $table->double('price')->default(0);
-            $table->integer('quantity')->default(0);
+            $table->double('quantity')->default(0);
             $table->tinyInteger('unit');
             $table->boolean('end_result')->default(false)->comment('При перерасчете суммы состава текущий состав добавляет в конечный итог');
             $table->timestamps();
